refactor: graphql codegen config
