+++ app/routes/($locale)._index.tsx
@@ -6,12 +6,18 @@ import {motion} from 'framer-motion';
 import {useWindowScroll} from 'react-use';
 
 import {ProductSwimlane, FeaturedCollections, Hero} from '~/components';
-import {MEDIA_FRAGMENT, PRODUCT_CARD_FRAGMENT} from '~/data/fragments';
 import {getHeroPlaceholder} from '~/lib/placeholders';
 import {seoPayload} from '~/lib/seo.server';
 import {routeHeaders} from '~/data/cache';
 import HeroCollection from '~/components/HeroCollection';
 import {useIsHomePath} from '~/lib/utils';
+import {
+  ALL_PAGES_QUERY,
+  COLLECTION_HERO_QUERY,
+  FEATURED_COLLECTIONS_QUERY,
+  HOMEPAGE_FEATURED_PRODUCTS_QUERY,
+  HOMEPAGE_SEO_QUERY,
+} from 'storefront-graphql';
 
 export const headers = routeHeaders;
 
@@ -193,130 +199,3 @@ export default function Homepage() {
     // </motion.div>
   );
 }
-
-const COLLECTION_CONTENT_FRAGMENT = `#graphql
-  fragment CollectionContent on Collection {
-    id
-    handle
-    title
-    descriptionHtml
-    heading: metafield(namespace: "hero", key: "title") {
-      value
-    }
-    byline: metafield(namespace: "hero", key: "byline") {
-      value
-    }
-    cta: metafield(namespace: "hero", key: "cta") {
-      value
-    }
-    spread: metafield(namespace: "hero", key: "spread") {
-      reference {
-        ...Media
-      }
-    }
-    spreadSecondary: metafield(namespace: "hero", key: "spread_secondary") {
-      reference {
-        ...Media
-      }
-    }
-  }
-  ${MEDIA_FRAGMENT}
-` as const;
-
-const HOMEPAGE_SEO_QUERY = `#graphql
-  query seoCollectionContent($handle: String, $country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    hero: collection(handle: $handle) {
-      ...CollectionContent
-    }
-    shop {
-      name
-      description
-    }
-  }
-  ${COLLECTION_CONTENT_FRAGMENT}
-` as const;
-
-const COLLECTION_HERO_QUERY = `#graphql
-  query heroCollectionContent($handle: String, $country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    hero: collection(handle: $handle) {
-      ...CollectionContent
-    }
-  }
-  ${COLLECTION_CONTENT_FRAGMENT}
-` as const;
-
-// @see: https://shopify.dev/api/storefront/current/queries/products
-export const HOMEPAGE_FEATURED_PRODUCTS_QUERY = `#graphql
-  query homepageFeaturedProducts($country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    products(first: 8) {
-      nodes {
-        ...ProductCard
-      }
-    }
-  }
-  ${PRODUCT_CARD_FRAGMENT}
-` as const;
-
-// @see: https://shopify.dev/api/storefront/current/queries/collections
-export const FEATURED_COLLECTIONS_QUERY = `#graphql
-  query homepageFeaturedCollections($country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    collections(
-      first: 4,
-      sortKey: UPDATED_AT
-    ) {
-      nodes {
-        id
-        title
-        handle
-        image {
-          altText
-          width
-          height
-          url
-        }
-      }
-    }
-  }
-` as const;
-
-export const ALL_PAGES_QUERY = `#graphql
-  query AllPages($country: CountryCode, $language: LanguageCode)
-  @inContext(country: $country, language: $language) {
-    pages(first: 100, query: "Homepage") {
-      nodes {
-        id
-        title
-        handle
-        pageTitle: metafield(namespace: "page", key: "title") {
-        value
-        }
-        imageCover: metafield(key: "image_cover", namespace: "page") {
-          reference {
-            ... on MediaImage {
-              id
-              image {
-                altText
-                height
-                id
-                url
-                width
-              }
-            }
-          }
-        }
-        collectionLink: metafield(key: "collection_link", namespace: "page") {
-          reference {
-            ... on Collection {
-              id
-              handle
-            }
-          }
-        }
-      }
-    }
-  }
-` as const;
