--- customer-accountapi.generated.d.ts
@@ -1,503 +1,0 @@
-/* eslint-disable eslint-comments/disable-enable-pair */
-/* eslint-disable eslint-comments/no-unlimited-disable */
-/* eslint-disable */
-import * as CustomerAccountAPI from '@shopify/hydrogen/customer-account-api-types';
-
-export type CustomerAddressUpdateMutationVariables = CustomerAccountAPI.Exact<{
-  address: CustomerAccountAPI.CustomerAddressInput;
-  addressId: CustomerAccountAPI.Scalars['ID']['input'];
-  defaultAddress?: CustomerAccountAPI.InputMaybe<
-    CustomerAccountAPI.Scalars['Boolean']['input']
-  >;
-}>;
-
-export type CustomerAddressUpdateMutation = {
-  customerAddressUpdate?: CustomerAccountAPI.Maybe<{
-    userErrors: Array<
-      Pick<
-        CustomerAccountAPI.UserErrorsCustomerAddressUserErrors,
-        'code' | 'field' | 'message'
-      >
-    >;
-  }>;
-};
-
-export type CustomerAddressDeleteMutationVariables = CustomerAccountAPI.Exact<{
-  addressId: CustomerAccountAPI.Scalars['ID']['input'];
-}>;
-
-export type CustomerAddressDeleteMutation = {
-  customerAddressDelete?: CustomerAccountAPI.Maybe<
-    Pick<
-      CustomerAccountAPI.CustomerAddressDeletePayload,
-      'deletedAddressId'
-    > & {
-      userErrors: Array<
-        Pick<
-          CustomerAccountAPI.UserErrorsCustomerAddressUserErrors,
-          'code' | 'field' | 'message'
-        >
-      >;
-    }
-  >;
-};
-
-export type CustomerAddressCreateMutationVariables = CustomerAccountAPI.Exact<{
-  address: CustomerAccountAPI.CustomerAddressInput;
-  defaultAddress?: CustomerAccountAPI.InputMaybe<
-    CustomerAccountAPI.Scalars['Boolean']['input']
-  >;
-}>;
-
-export type CustomerAddressCreateMutation = {
-  customerAddressCreate?: CustomerAccountAPI.Maybe<{
-    customerAddress?: CustomerAccountAPI.Maybe<
-      Pick<CustomerAccountAPI.CustomerAddress, 'id'>
-    >;
-    userErrors: Array<
-      Pick<
-        CustomerAccountAPI.UserErrorsCustomerAddressUserErrors,
-        'code' | 'field' | 'message'
-      >
-    >;
-  }>;
-};
-
-export type OrderCardFragment = Pick<
-  CustomerAccountAPI.Order,
-  'id' | 'number' | 'processedAt' | 'financialStatus'
-> & {
-  fulfillments: {nodes: Array<Pick<CustomerAccountAPI.Fulfillment, 'status'>>};
-  totalPrice: Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>;
-  lineItems: {
-    edges: Array<{
-      node: Pick<CustomerAccountAPI.LineItem, 'title'> & {
-        image?: CustomerAccountAPI.Maybe<
-          Pick<CustomerAccountAPI.Image, 'altText' | 'height' | 'url' | 'width'>
-        >;
-      };
-    }>;
-  };
-};
-
-export type AddressPartialFragment = Pick<
-  CustomerAccountAPI.CustomerAddress,
-  | 'id'
-  | 'formatted'
-  | 'firstName'
-  | 'lastName'
-  | 'company'
-  | 'address1'
-  | 'address2'
-  | 'territoryCode'
-  | 'zoneCode'
-  | 'city'
-  | 'zip'
-  | 'phoneNumber'
->;
-
-export type CustomerDetailsFragment = Pick<
-  CustomerAccountAPI.Customer,
-  'firstName' | 'lastName'
-> & {
-  phoneNumber?: CustomerAccountAPI.Maybe<
-    Pick<CustomerAccountAPI.CustomerPhoneNumber, 'phoneNumber'>
-  >;
-  emailAddress?: CustomerAccountAPI.Maybe<
-    Pick<CustomerAccountAPI.CustomerEmailAddress, 'emailAddress'>
-  >;
-  defaultAddress?: CustomerAccountAPI.Maybe<
-    Pick<
-      CustomerAccountAPI.CustomerAddress,
-      | 'id'
-      | 'formatted'
-      | 'firstName'
-      | 'lastName'
-      | 'company'
-      | 'address1'
-      | 'address2'
-      | 'territoryCode'
-      | 'zoneCode'
-      | 'city'
-      | 'zip'
-      | 'phoneNumber'
-    >
-  >;
-  addresses: {
-    edges: Array<{
-      node: Pick<
-        CustomerAccountAPI.CustomerAddress,
-        | 'id'
-        | 'formatted'
-        | 'firstName'
-        | 'lastName'
-        | 'company'
-        | 'address1'
-        | 'address2'
-        | 'territoryCode'
-        | 'zoneCode'
-        | 'city'
-        | 'zip'
-        | 'phoneNumber'
-      >;
-    }>;
-  };
-  orders: {
-    edges: Array<{
-      node: Pick<
-        CustomerAccountAPI.Order,
-        'id' | 'number' | 'processedAt' | 'financialStatus'
-      > & {
-        fulfillments: {
-          nodes: Array<Pick<CustomerAccountAPI.Fulfillment, 'status'>>;
-        };
-        totalPrice: Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>;
-        lineItems: {
-          edges: Array<{
-            node: Pick<CustomerAccountAPI.LineItem, 'title'> & {
-              image?: CustomerAccountAPI.Maybe<
-                Pick<
-                  CustomerAccountAPI.Image,
-                  'altText' | 'height' | 'url' | 'width'
-                >
-              >;
-            };
-          }>;
-        };
-      };
-    }>;
-  };
-};
-
-export type CustomerDetailsQueryVariables = CustomerAccountAPI.Exact<{
-  [key: string]: never;
-}>;
-
-export type CustomerDetailsQuery = {
-  customer: Pick<CustomerAccountAPI.Customer, 'firstName' | 'lastName'> & {
-    phoneNumber?: CustomerAccountAPI.Maybe<
-      Pick<CustomerAccountAPI.CustomerPhoneNumber, 'phoneNumber'>
-    >;
-    emailAddress?: CustomerAccountAPI.Maybe<
-      Pick<CustomerAccountAPI.CustomerEmailAddress, 'emailAddress'>
-    >;
-    defaultAddress?: CustomerAccountAPI.Maybe<
-      Pick<
-        CustomerAccountAPI.CustomerAddress,
-        | 'id'
-        | 'formatted'
-        | 'firstName'
-        | 'lastName'
-        | 'company'
-        | 'address1'
-        | 'address2'
-        | 'territoryCode'
-        | 'zoneCode'
-        | 'city'
-        | 'zip'
-        | 'phoneNumber'
-      >
-    >;
-    addresses: {
-      edges: Array<{
-        node: Pick<
-          CustomerAccountAPI.CustomerAddress,
-          | 'id'
-          | 'formatted'
-          | 'firstName'
-          | 'lastName'
-          | 'company'
-          | 'address1'
-          | 'address2'
-          | 'territoryCode'
-          | 'zoneCode'
-          | 'city'
-          | 'zip'
-          | 'phoneNumber'
-        >;
-      }>;
-    };
-    orders: {
-      edges: Array<{
-        node: Pick<
-          CustomerAccountAPI.Order,
-          'id' | 'number' | 'processedAt' | 'financialStatus'
-        > & {
-          fulfillments: {
-            nodes: Array<Pick<CustomerAccountAPI.Fulfillment, 'status'>>;
-          };
-          totalPrice: Pick<
-            CustomerAccountAPI.MoneyV2,
-            'amount' | 'currencyCode'
-          >;
-          lineItems: {
-            edges: Array<{
-              node: Pick<CustomerAccountAPI.LineItem, 'title'> & {
-                image?: CustomerAccountAPI.Maybe<
-                  Pick<
-                    CustomerAccountAPI.Image,
-                    'altText' | 'height' | 'url' | 'width'
-                  >
-                >;
-              };
-            }>;
-          };
-        };
-      }>;
-    };
-  };
-};
-
-export type OrderMoneyFragment = Pick<
-  CustomerAccountAPI.MoneyV2,
-  'amount' | 'currencyCode'
->;
-
-export type DiscountApplicationFragment = {
-  value:
-    | ({__typename: 'MoneyV2'} & Pick<
-        CustomerAccountAPI.MoneyV2,
-        'amount' | 'currencyCode'
-      >)
-    | ({__typename: 'PricingPercentageValue'} & Pick<
-        CustomerAccountAPI.PricingPercentageValue,
-        'percentage'
-      >);
-};
-
-export type OrderLineItemFullFragment = Pick<
-  CustomerAccountAPI.LineItem,
-  'id' | 'title' | 'quantity' | 'variantTitle'
-> & {
-  price?: CustomerAccountAPI.Maybe<
-    Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-  >;
-  discountAllocations: Array<{
-    allocatedAmount: Pick<
-      CustomerAccountAPI.MoneyV2,
-      'amount' | 'currencyCode'
-    >;
-    discountApplication: {
-      value:
-        | ({__typename: 'MoneyV2'} & Pick<
-            CustomerAccountAPI.MoneyV2,
-            'amount' | 'currencyCode'
-          >)
-        | ({__typename: 'PricingPercentageValue'} & Pick<
-            CustomerAccountAPI.PricingPercentageValue,
-            'percentage'
-          >);
-    };
-  }>;
-  totalDiscount: Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>;
-  image?: CustomerAccountAPI.Maybe<
-    Pick<
-      CustomerAccountAPI.Image,
-      'altText' | 'height' | 'url' | 'id' | 'width'
-    >
-  >;
-};
-
-export type OrderFragment = Pick<
-  CustomerAccountAPI.Order,
-  'id' | 'name' | 'statusPageUrl' | 'processedAt'
-> & {
-  fulfillments: {nodes: Array<Pick<CustomerAccountAPI.Fulfillment, 'status'>>};
-  totalTax?: CustomerAccountAPI.Maybe<
-    Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-  >;
-  totalPrice: Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>;
-  subtotal?: CustomerAccountAPI.Maybe<
-    Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-  >;
-  shippingAddress?: CustomerAccountAPI.Maybe<
-    Pick<
-      CustomerAccountAPI.CustomerAddress,
-      'name' | 'formatted' | 'formattedArea'
-    >
-  >;
-  discountApplications: {
-    nodes: Array<{
-      value:
-        | ({__typename: 'MoneyV2'} & Pick<
-            CustomerAccountAPI.MoneyV2,
-            'amount' | 'currencyCode'
-          >)
-        | ({__typename: 'PricingPercentageValue'} & Pick<
-            CustomerAccountAPI.PricingPercentageValue,
-            'percentage'
-          >);
-    }>;
-  };
-  lineItems: {
-    nodes: Array<
-      Pick<
-        CustomerAccountAPI.LineItem,
-        'id' | 'title' | 'quantity' | 'variantTitle'
-      > & {
-        price?: CustomerAccountAPI.Maybe<
-          Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-        >;
-        discountAllocations: Array<{
-          allocatedAmount: Pick<
-            CustomerAccountAPI.MoneyV2,
-            'amount' | 'currencyCode'
-          >;
-          discountApplication: {
-            value:
-              | ({__typename: 'MoneyV2'} & Pick<
-                  CustomerAccountAPI.MoneyV2,
-                  'amount' | 'currencyCode'
-                >)
-              | ({__typename: 'PricingPercentageValue'} & Pick<
-                  CustomerAccountAPI.PricingPercentageValue,
-                  'percentage'
-                >);
-          };
-        }>;
-        totalDiscount: Pick<
-          CustomerAccountAPI.MoneyV2,
-          'amount' | 'currencyCode'
-        >;
-        image?: CustomerAccountAPI.Maybe<
-          Pick<
-            CustomerAccountAPI.Image,
-            'altText' | 'height' | 'url' | 'id' | 'width'
-          >
-        >;
-      }
-    >;
-  };
-};
-
-export type OrderQueryVariables = CustomerAccountAPI.Exact<{
-  orderId: CustomerAccountAPI.Scalars['ID']['input'];
-}>;
-
-export type OrderQuery = {
-  order?: CustomerAccountAPI.Maybe<
-    Pick<
-      CustomerAccountAPI.Order,
-      'id' | 'name' | 'statusPageUrl' | 'processedAt'
-    > & {
-      fulfillments: {
-        nodes: Array<Pick<CustomerAccountAPI.Fulfillment, 'status'>>;
-      };
-      totalTax?: CustomerAccountAPI.Maybe<
-        Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-      >;
-      totalPrice: Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>;
-      subtotal?: CustomerAccountAPI.Maybe<
-        Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-      >;
-      shippingAddress?: CustomerAccountAPI.Maybe<
-        Pick<
-          CustomerAccountAPI.CustomerAddress,
-          'name' | 'formatted' | 'formattedArea'
-        >
-      >;
-      discountApplications: {
-        nodes: Array<{
-          value:
-            | ({__typename: 'MoneyV2'} & Pick<
-                CustomerAccountAPI.MoneyV2,
-                'amount' | 'currencyCode'
-              >)
-            | ({__typename: 'PricingPercentageValue'} & Pick<
-                CustomerAccountAPI.PricingPercentageValue,
-                'percentage'
-              >);
-        }>;
-      };
-      lineItems: {
-        nodes: Array<
-          Pick<
-            CustomerAccountAPI.LineItem,
-            'id' | 'title' | 'quantity' | 'variantTitle'
-          > & {
-            price?: CustomerAccountAPI.Maybe<
-              Pick<CustomerAccountAPI.MoneyV2, 'amount' | 'currencyCode'>
-            >;
-            discountAllocations: Array<{
-              allocatedAmount: Pick<
-                CustomerAccountAPI.MoneyV2,
-                'amount' | 'currencyCode'
-              >;
-              discountApplication: {
-                value:
-                  | ({__typename: 'MoneyV2'} & Pick<
-                      CustomerAccountAPI.MoneyV2,
-                      'amount' | 'currencyCode'
-                    >)
-                  | ({__typename: 'PricingPercentageValue'} & Pick<
-                      CustomerAccountAPI.PricingPercentageValue,
-                      'percentage'
-                    >);
-              };
-            }>;
-            totalDiscount: Pick<
-              CustomerAccountAPI.MoneyV2,
-              'amount' | 'currencyCode'
-            >;
-            image?: CustomerAccountAPI.Maybe<
-              Pick<
-                CustomerAccountAPI.Image,
-                'altText' | 'height' | 'url' | 'id' | 'width'
-              >
-            >;
-          }
-        >;
-      };
-    }
-  >;
-};
-
-export type CustomerUpdateMutationVariables = CustomerAccountAPI.Exact<{
-  customer: CustomerAccountAPI.CustomerUpdateInput;
-}>;
-
-export type CustomerUpdateMutation = {
-  customerUpdate?: CustomerAccountAPI.Maybe<{
-    userErrors: Array<
-      Pick<
-        CustomerAccountAPI.UserErrorsCustomerUserErrors,
-        'code' | 'field' | 'message'
-      >
-    >;
-  }>;
-};
-
-interface GeneratedQueryTypes {
-  '#graphql\n  query CustomerDetails {\n    customer {\n      ...CustomerDetails\n    }\n  }\n  #graphql\n  fragment OrderCard on Order {\n    id\n    number\n    processedAt\n    financialStatus\n    fulfillments(first: 1) {\n      nodes {\n        status\n      }\n    }\n    totalPrice {\n      amount\n      currencyCode\n    }\n    lineItems(first: 2) {\n      edges {\n        node {\n          title\n          image {\n            altText\n            height\n            url\n            width\n          }\n        }\n      }\n    }\n  }\n\n  fragment AddressPartial on CustomerAddress {\n    id\n    formatted\n    firstName\n    lastName\n    company\n    address1\n    address2\n    territoryCode\n    zoneCode\n    city\n    zip\n    phoneNumber\n  }\n\n  fragment CustomerDetails on Customer {\n    firstName\n    lastName\n    phoneNumber {\n      phoneNumber\n    }\n    emailAddress {\n      emailAddress\n    }\n    defaultAddress {\n      ...AddressPartial\n    }\n    addresses(first: 6) {\n      edges {\n        node {\n          ...AddressPartial\n        }\n      }\n    }\n    orders(first: 250, sortKey: PROCESSED_AT, reverse: true) {\n      edges {\n        node {\n          ...OrderCard\n        }\n      }\n    }\n  }\n\n': {
-    return: CustomerDetailsQuery;
-    variables: CustomerDetailsQueryVariables;
-  };
-  '#graphql\n  fragment OrderMoney on MoneyV2 {\n    amount\n    currencyCode\n  }\n  fragment DiscountApplication on DiscountApplication {\n    value {\n      __typename\n      ... on MoneyV2 {\n        ...OrderMoney\n      }\n      ... on PricingPercentageValue {\n        percentage\n      }\n    }\n  }\n  fragment OrderLineItemFull on LineItem {\n    id\n    title\n    quantity\n    price {\n      ...OrderMoney\n    }\n    discountAllocations {\n      allocatedAmount {\n        ...OrderMoney\n      }\n      discountApplication {\n        ...DiscountApplication\n      }\n    }\n    totalDiscount {\n      ...OrderMoney\n    }\n    image {\n      altText\n      height\n      url\n      id\n      width\n    }\n    variantTitle\n  }\n  fragment Order on Order {\n    id\n    name\n    statusPageUrl\n    processedAt\n    fulfillments(first: 1) {\n      nodes {\n        status\n      }\n    }\n    totalTax {\n      ...OrderMoney\n    }\n    totalPrice {\n      ...OrderMoney\n    }\n    subtotal {\n      ...OrderMoney\n    }\n    shippingAddress {\n      name\n      formatted(withName: true)\n      formattedArea\n    }\n    discountApplications(first: 100) {\n      nodes {\n        ...DiscountApplication\n      }\n    }\n    lineItems(first: 100) {\n      nodes {\n        ...OrderLineItemFull\n      }\n    }\n  }\n  query Order($orderId: ID!) {\n    order(id: $orderId) {\n      ... on Order {\n        ...Order\n      }\n    }\n  }\n': {
-    return: OrderQuery;
-    variables: OrderQueryVariables;
-  };
-}
-
-interface GeneratedMutationTypes {
-  '#graphql\n  mutation customerAddressUpdate(\n    $address: CustomerAddressInput!\n    $addressId: ID!\n    $defaultAddress: Boolean\n ) {\n    customerAddressUpdate(\n      address: $address\n      addressId: $addressId\n      defaultAddress: $defaultAddress\n    ) {\n      userErrors {\n        code\n        field\n        message\n      }\n    }\n  }\n': {
-    return: CustomerAddressUpdateMutation;
-    variables: CustomerAddressUpdateMutationVariables;
-  };
-  '#graphql\n  mutation customerAddressDelete(\n    $addressId: ID!,\n  ) {\n    customerAddressDelete(addressId: $addressId) {\n      deletedAddressId\n      userErrors {\n        code\n        field\n        message\n      }\n    }\n  }\n': {
-    return: CustomerAddressDeleteMutation;
-    variables: CustomerAddressDeleteMutationVariables;
-  };
-  '#graphql\n  mutation customerAddressCreate(\n    $address: CustomerAddressInput!\n    $defaultAddress: Boolean\n  ) {\n    customerAddressCreate(\n      address: $address\n      defaultAddress: $defaultAddress\n    ) {\n      customerAddress {\n        id\n      }\n      userErrors {\n        code\n        field\n        message\n      }\n    }\n  }\n': {
-    return: CustomerAddressCreateMutation;
-    variables: CustomerAddressCreateMutationVariables;
-  };
-  '#graphql\nmutation customerUpdate($customer: CustomerUpdateInput!) {\n  customerUpdate(input: $customer) {\n    userErrors {\n      code\n      field\n      message\n    }\n  }\n}\n': {
-    return: CustomerUpdateMutation;
-    variables: CustomerUpdateMutationVariables;
-  };
-}
-
-declare module '@shopify/hydrogen' {
-  interface CustomerAccountQueries extends GeneratedQueryTypes {}
-  interface CustomerAccountMutations extends GeneratedMutationTypes {}
-}
