+++ storefront-graphql.ts
@@ -1,0 +1,217 @@
+export const MEDIA_FRAGMENT = `#graphql
+  fragment Media on Media {
+    __typename
+    mediaContentType
+    alt
+    previewImage {
+      url
+    }
+    ... on MediaImage {
+      id
+      image {
+        id
+        url
+        width
+        height
+      }
+    }
+    ... on Video {
+      id
+      sources {
+        mimeType
+        url
+      }
+    }
+    ... on Model3d {
+      id
+      sources {
+        mimeType
+        url
+      }
+    }
+    ... on ExternalVideo {
+      id
+      embedUrl
+      host
+    }
+  }
+`;
+
+export const PRODUCT_CARD_FRAGMENT = `#graphql
+  fragment ProductCard on Product {
+    id
+    title
+    publishedAt
+    handle
+    vendor
+    variants(first: 1) {
+      nodes {
+        id
+        availableForSale
+        image {
+          url
+          altText
+          width
+          height
+        }
+        price {
+          amount
+          currencyCode
+        }
+        compareAtPrice {
+          amount
+          currencyCode
+        }
+        selectedOptions {
+          name
+          value
+        }
+        product {
+          handle
+          title
+        }
+      }
+    }
+  }
+`;
+
+export const FEATURED_COLLECTION_FRAGMENT = `#graphql
+  fragment FeaturedCollectionDetails on Collection {
+    id
+    title
+    handle
+    image {
+      altText
+      width
+      height
+      url
+    }
+  }
+`;
+
+export const COLLECTION_CONTENT_FRAGMENT = `#graphql
+  fragment CollectionContent on Collection {
+    id
+    handle
+    title
+    descriptionHtml
+    heading: metafield(namespace: "hero", key: "title") {
+      value
+    }
+    byline: metafield(namespace: "hero", key: "byline") {
+      value
+    }
+    cta: metafield(namespace: "hero", key: "cta") {
+      value
+    }
+    spread: metafield(namespace: "hero", key: "spread") {
+      reference {
+        ...Media
+      }
+    }
+    spreadSecondary: metafield(namespace: "hero", key: "spread_secondary") {
+      reference {
+        ...Media
+      }
+    }
+  }
+  ${MEDIA_FRAGMENT}
+` as const;
+
+export const HOMEPAGE_SEO_QUERY = `#graphql
+  query seoCollectionContent($handle: String, $country: CountryCode, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    hero: collection(handle: $handle) {
+      ...CollectionContent
+    }
+    shop {
+      name
+      description
+    }
+  }
+  ${COLLECTION_CONTENT_FRAGMENT}
+` as const;
+
+export const COLLECTION_HERO_QUERY = `#graphql
+  query heroCollectionContent($handle: String, $country: CountryCode, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    hero: collection(handle: $handle) {
+      ...CollectionContent
+    }
+  }
+  ${COLLECTION_CONTENT_FRAGMENT}
+` as const;
+
+// @see: https://shopify.dev/api/storefront/current/queries/products
+export const HOMEPAGE_FEATURED_PRODUCTS_QUERY = `#graphql
+  query homepageFeaturedProducts($country: CountryCode, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    products(first: 8) {
+      nodes {
+        ...ProductCard
+      }
+    }
+  }
+  ${PRODUCT_CARD_FRAGMENT}
+` as const;
+
+// @see: https://shopify.dev/api/storefront/current/queries/collections
+export const FEATURED_COLLECTIONS_QUERY = `#graphql
+  query homepageFeaturedCollections($country: CountryCode, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    collections(
+      first: 4,
+      sortKey: UPDATED_AT
+    ) {
+      nodes {
+        id
+        title
+        handle
+        image {
+          altText
+          width
+          height
+          url
+        }
+      }
+    }
+  }
+` as const;
+
+export const ALL_PAGES_QUERY = `#graphql
+  query AllPages($country: CountryCode, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    pages(first: 100, query: "Homepage") {
+      nodes {
+        id
+        title
+        handle
+        pageTitle: metafield(namespace: "page", key: "title") {
+        value
+        }
+        imageCover: metafield(key: "image_cover", namespace: "page") {
+          reference {
+            ... on MediaImage {
+              id
+              image {
+                altText
+                height
+                id
+                url
+                width
+              }
+            }
+          }
+        }
+        collectionLink: metafield(key: "collection_link", namespace: "page") {
+          reference {
+            ... on Collection {
+              id
+              handle
+            }
+          }
+        }
+      }
+    }
+  }
+` as const;
